--- crates/deskulpt-core/src/apis/mod.rs
@@ -1,3 +1,0 @@
-//! Deskulpt APIs (🚧 to be refactored 🚧).
-
-pub mod sys;
