+++ crates/deskulpt-plugin-sys/README.md
@@ -1,0 +1,3 @@
+This crate implements the [Deskulpt](https://csci-shu-410-se-project.github.io/Deskulpt/) system information plugin.
+
+⚠️ This crate is meant to be consumed as a Deskulpt plugin instead of as a library. Private items are documented for reference of plugin authors and Deskulpt developers.
