+++ crates/deskulpt-plugin-sys/src/commands/mod.rs
@@ -1,0 +1,6 @@
+//! System information plugin commands.
+
+mod get_system_info;
+
+#[doc(hidden)]
+pub use get_system_info::GetSystemInfo;
